test: extracted Form Search Tags tests into seperate test file as testing seperate component
