+++ client/techlog/src/__tests__/Form.test.js
@@ -1,14 +1,16 @@
 import React from 'react';
-import NewReport from '../Components/NewReport/NewReport';
 import Form from '../Form/Form';
 import { MemoryRouter } from 'react-router-dom';
 import { screen, getByLabelText, render, fireEvent, getByRole, waitFor, getByText } from '@testing-library/react';
 import '@testing-library/jest-dom';
 
-describe('Form input', () => {
+beforeEach(() => {
+  render(<MemoryRouter><Form /></MemoryRouter>);
+})
+
+describe('Form', () => {
   
   it('Should pesist input text when element not in focus', () => {
-    render(<MemoryRouter><Form /></MemoryRouter>);
     const titleInput = screen.getByLabelText(/report title/i);
     const descriptionInput = screen.getByLabelText(/description/i);
     // input text into field
